Refactor CBOR encoder/decoder module structure (#664)
